feat(version): add version command to display project version
