+++ .vscode/settings.json
@@ -8,6 +8,7 @@
     "editor.formatOnSave": true,
     "editor.codeActionsOnSave": {
       "source.fixAll": "explicit",
+      "source.organizeImports": "explicit",
       "notebook.source.organizeImports": "explicit"
     },
     "editor.defaultFormatter": "charliermarsh.ruff"
