+++ src/foxy_project/changelog/templates/compact.jinja2
