+++ src/foxy_project/changelog/templates/lastrelease.jinja2
