+++ src/foxy_project/changelog/templates/macros.jinja2
