+++ src/foxy_project/version/config/setuptools.py
